share testhelper
